// src/app/core/model/major/model/Major.ts
export interface Major {
  idMajor: number;
  name: string;
  description: string;
}
export interface MajorCreateDto {
  name: string;
  description: string;
}
export interface UpdateMajorDto {
  idMajor: number;
  name: string;
  description: string;
}
export interface DeleteMajorDto {
  idMajor: number;
  name: string;
  description: string;
}